==> 02-lgbsfeature.ipynb <==
# General imports
import numpy as np
import pandas as pd
import os, sys, gc, time, warnings, pickle, psutil, random
import time

warnings.filterwarnings('ignore')

def get_memory_usage():
    return np.round(psutil.Process(os.getpid()).memory_info()[0]/2.**30, 2) 
        
def sizeof_fmt(num, suffix='B'):
    for unit in ['','Ki','Mi','Gi','Ti','Pi','Ei','Zi']:
        if abs(num) < 1024.0:
            return "%3.1f%s%s" % (num, unit, suffix)
        num /= 1024.0
    return "%.1f%s%s" % (num, 'Yi', suffix)

########################### Vars
#################################################################################
TARGET = 'sales'         # Our main target
END_TRAIN = 1913         # Last day in train set
MAIN_INDEX = ['id','d']  # We can identify item by these columns

#################################################################################
print('Load Main Data')

# We will need only train dataset
# to show lags concept
train_df = pd.read_csv('./data/sales_train_validation.csv')

# To make all calculations faster
# we will limit dataset by 'CA' state
train_df = train_df[train_df['state_id']=='CA']

########################### Data Representation
#################################################################################

# Let's check our shape
print('Shape', train_df.shape)

index_columns = ['id','item_id','dept_id','cat_id','store_id','state_id']
train_df = pd.melt(train_df, 
                  id_vars = index_columns, 
                  var_name = 'd', 
                  value_name = TARGET)

train_df[train_df['id']=='HOBBIES_1_001_CA_1_validation'].iloc[:10]

train_df['day'] = train_df['d'].apply(lambda x: x[2:]).astype(np.int16)
train_df = train_df[train_df['day']>703]
print(train_df.shape) 
train_df = train_df.iloc[:1000, :]
train_df.head()

train_df['d'] = train_df['d'].apply(lambda x: x[2:]).astype(np.int16)

icols = ['id','item_id','dept_id','cat_id','store_id','state_id']
for col in icols:
    train_df[col] = train_df[col].astype('category')

########################### Lags creation
#################################################################################

temp_df = train_df[['id', 'd', TARGET]]

start_time = time.time()
for i in range(1, 8):
    print('Shifting:', i)
    temp_df['lag_' + str(i)] = temp_df.groupby(
        ['id'])[TARGET].transform(lambda x: x.shift(i))

print('%0.2f min: Time for loops' % ((time.time() - start_time) / 60))

# Or same in "compact" manner
LAG_DAYS = [col for col in range(1, 8)]
temp_df = train_df[['id', 'd', TARGET]]

start_time = time.time()
temp_df = temp_df.assign(
    **{
        '{}_lag_{}'.format(col, l): temp_df.groupby(['id'])[col].transform(
            lambda x: x.shift(l))
        for l in LAG_DAYS for col in [TARGET]
    })

print('%0.2f min: Time for bulk shift' % ((time.time() - start_time) / 60))

temp_df[temp_df['id']=='HOBBIES_1_001_CA_1_validation'].iloc[:10]

########################### Rolling lags
#################################################################################

temp_df = train_df[['id', 'd', 'sales']]

start_time = time.time()
for i in [14, 30, 60]:
    print('Rolling period:', i)
    temp_df['rolling_mean_' + str(i)] = temp_df.groupby(
        ['id'])[TARGET].transform(lambda x: x.shift(1).rolling(i).mean())
    temp_df['rolling_std_' + str(i)] = temp_df.groupby(
        ['id'])[TARGET].transform(lambda x: x.shift(1).rolling(i).std())

print('%0.2f min: Time for loop' % ((time.time() - start_time) / 60))

# The result
temp_df[temp_df['id'] == 'HOBBIES_1_002_CA_1_validation'].iloc[:20]

# Same for NaNs values - it's normal
# because there is no data for
# 0*(rolling_period),-1*(rolling_period),-2*(rolling_period)

########################### Memory ussage
#################################################################################
print("{:>20}: {:>8}".format(
    'Original rolling df', sizeof_fmt(temp_df.memory_usage(index=True).sum())))

temp_df = temp_df.iloc[:, 3:]
print("{:>20}: {:>8}".format(
    'Values rolling df', sizeof_fmt(temp_df.memory_usage(index=True).sum())))

from scipy import sparse
temp_matrix = sparse.csr_matrix(temp_df)

# restore to df
temp_matrix_restored = pd.DataFrame(temp_matrix.todense())
restored_cols = ['roll_' + str(i) for i in list(temp_matrix_restored)]
temp_matrix_restored.columns = restored_cols

########################### Remove old objects
#################################################################################
del temp_df, train_df, temp_matrix, temp_matrix_restored

grid_df = pd.read_pickle('./data/output/grid_part_1.pkl')

# We need only 'id','d','sales'
# to make lags and rollings
grid_df = grid_df[['id', 'd', 'sales']]
SHIFT_DAY = 28

start_time = time.time()
print('Create lags')

LAG_DAYS = [col for col in range(SHIFT_DAY, SHIFT_DAY + 15)]
grid_df = grid_df.assign(
    **{
        '{}_lag_{}'.format(col, l): grid_df.groupby(['id'])[col].transform(
            lambda x: x.shift(l))
        for l in LAG_DAYS for col in [TARGET]
    })

for col in list(grid_df):
    if 'lag' in col:
        grid_df[col] = grid_df[col].astype(np.float16)

print('%0.2f min: Lags' % ((time.time() - start_time) / 60))

start_time = time.time()
print('Create rolling aggs')

for i in [7, 14, 30, 60, 180]:
    print('Rolling period:', i)
    grid_df['rolling_mean_' + str(i)] = grid_df.groupby(
        ['id'])[TARGET].transform(
            lambda x: x.shift(SHIFT_DAY).rolling(i).mean()).astype(np.float16)
    grid_df['rolling_std_' + str(i)] = grid_df.groupby([
        'id'
    ])[TARGET].transform(lambda x: x.shift(SHIFT_DAY).rolling(i).std()).astype(
        np.float16)

# Rollings
# with sliding shift
for d_shift in [1, 7, 14]:
    print('Shifting period:', d_shift)
    for d_window in [7, 14, 30, 60]:
        col_name = 'rolling_mean_tmp_' + str(d_shift) + '_' + str(d_window)
        grid_df[col_name] = grid_df.groupby(['id'])[TARGET].transform(
            lambda x: x.shift(d_shift).rolling(d_window).mean()).astype(
                np.float16)

print('%0.2f min: Lags' % ((time.time() - start_time) / 60))

#################################################################################
print('Save lags and rollings')
grid_df.to_pickle('./data/output/lags_df_' + str(SHIFT_DAY) + '.pkl')
